# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Order Management System"
tags:
- name: "products"
  description: "Access to Inventory Management System"
schemes:
- "http"
paths:
  /products/{id}:
    get:
      tags:
      - "products"
      summary: "Fetch Product based on product id"
      description: "Returns a product based on product id"
      operationId: "getProducts"
      produces:
      - "application/json"
      parameters:
        - name: id
          in: path
          type: string
          description: The id of the product.
          required: true
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Products"
        400:
          description: "Bad request"
        404:
          description: "Not Found"      
        500:
          description: "Internal Server Error"
      security:
      - api_key: []
  /products:
    post:
      tags:
      - "products"
      summary: "Add product(s) to inventory"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "adding product(s) to the inventory"
        required: true
        schema:
          $ref: "#/definitions/Products"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Products"
        400:
          description: "Bad request"        
        500:
          description: "Internal Server Error"
  /products/order:
    post:
      tags:
      - "products"
      summary: "Order products"
      description: "Order products"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "placing orders"
        required: true
        schema:
          $ref: "#/definitions/ProductIDS"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Products"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Order not found"
        500:
          description: "Internal Server Error"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost:8080/oauth/dialog"
    flow: "implicit"
    scopes:
      write:products: "add products in your account"
      read:products: "read products in your account"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Products:
    type: "object"
    properties:
      product_type:
        type: "string"
        description: "Product type"
        example: mobile
      seller_id:
        type: "string"
        description: "Seller ID"
        example: XYZ_SELLER
      size:
        type: "integer"
        format: int32
        description: "Size"
        example: 5
      brand:
        type: "string"
        description: "Brand"
        example: nokia
      metadata:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
              description: name of the object
        example:  [{ "color": "blue", "items": 2, "price": "30,000" }, { "color": "black", "items": 1, "price": "35,000" }]
        # required: true
      location:
        type: "string"
        description: "Location"
        example: Tamilnadu
      location_id:
        type: "string"
        description: "Location_id"
        example: CHE_006
  ProductIDS:
    type: "object"
    properties:
      product_ids:
        type: "array"
        description: "Product IDS"
        example: ["1","2"]
        items:
            properties:
              type:
                type: string
                example: "1" 
      address:
        type: "string"
        description: "Delivery Address"
        example: "No:1, Radission Apartments, Anna nagar, Chennai-001"             
      price:
        type: "string"
        description: "Order Price"
        example: "45,000"               